# server/src/interfaces/http/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a socco project",
    "version": "1.0.0",
    "title": "SOCCO Backend",
    "termsOfService": "",
    "contact": {
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
  "paths": {
    "/product/": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Get all the products",
        "description": "Return the list of products",
        "operationId": "findAll",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number of the result. Default value is 1.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "number of products in the result. Default value is 10.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/product/findBySearchText": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Finds Products by Search term",
        "description": "Return the list of products matching the search",
        "operationId": "findBySearchText",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "searchText",
            "in": "query",
            "description": "Status values that need to be considered for filter",
            "required": true,
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number of the result. Default value is 1.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "number of products in the result. Default value is 10.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/product/findByCategory": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Finds Products by category",
        "description": "Find Products by category.",
        "operationId": "findByCategory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Category Id of the products",
            "required": true,
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          },
          "400": {
            "description": "Category Id"
          }
        }
      }
    },
    "/product/findByCommunity": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Finds Products by community",
        "description": "Find Products by community.",
        "operationId": "findByCommunity",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "community",
            "in": "query",
            "description": "Community Id",
            "required": true,
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          },
          "400": {
            "description": "Community Id"
          }
        }
      }
    },
    "/product/findByCategoryAndCommunity": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Finds Products by category and community",
        "description": "Find Products by category and community.",
        "operationId": "findByCategoryAndCommunity",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "community",
            "in": "query",
            "description": "Community Id",
            "required": true,
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "name": "category",
            "in": "query",
            "description": "Category Id",
            "required": true,
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          },
          "400": {
            "description": "Community/Category Id"
          }
        }
      }
    },
    "/product/{productId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Get Product Specific to the Id",
        "description": "Returns the details of the product",
        "operationId": "findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of the product",
            "required": true,
            "type": "integer",
            "maximum": 10,
            "minimum": 1,
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Product Id"
          },
          "404": {
            "description": "Product Not Found"
          }
        }
      }
    },
    "/product/getAttributes/{productId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Get Product Specific to the Id",
        "description": "Returns the details of the product",
        "operationId": "findAttributes",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of the product",
            "required": true,
            "type": "integer",
            "maximum": 10,
            "minimum": 1,
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ProductAttributes"
            }
          },
          "400": {
            "description": "Product Id"
          },
          "404": {
            "description": "Product Not Found"
          }
        }
      }
    },
...
    "/community/findBySearchText": {
      "get": {
        "tags": [
          "community"
        ],
        "summary": "Finds Community by Search term",
        "description": "Return the list of communities matching the search",
        "operationId": "findBySearchText",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "searchText",
            "in": "query",
            "description": "Status values that need to be considered for filter",
            "required": true,
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number of the result. Default value is 1.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "number of products in the result. Default value is 10.",
            "required": false,
            "type": "number",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Community"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Products": {
      "type": "object",
      "required": [
        "products",
        "pageCount",
        "itemCount",
        "limit",
        "currentPage"
      ],
      "properties": {
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        },
        "pageCount": {
          "type": "number"
        },
        "itemCount": {
          "type": "number"
        },
        "limit": {
          "type": "number"
        },
        "currentPage": {
          "type": "number"
        }
      }
    },
    "Product": {
      "type": "object",
      "required": [
        "name",
        "description",
        "price",
        "discountedPrice",
        "display"
      ],
      "properties": {
        "productId": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "example": "Chartres Cathedral"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "discountedPrice": {
          "type": "number"
        },
        "image": {
          "type": "string"
        },
        "image2": {
          "type": "string"
        },
        "thumbnail": {
          "type": "string"
        },
        "display": {
          "type": "integer"
        }
      }
    },
    "ProductAttributes": {
      "type": "object",
      "required": [
        "attributes"
      ],
      "properties": {
        "attributes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductAttribute"
          }
        }
      }
    },
    "ProductAttribute": {
      "type": "object",
      "required": [
        "name",
        "attributeId",
        "attributeValue"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "Size"
        },
        "attributeId": {
          "type": "number"
        },
        "attributeValue": {
          "type": "string"
        }
      }
    },
    "Community": {
      "type": "object",
      "properties": {
        "agent_id": {
          "type": "integer",
          "format": "int64"
        },
        "community_name": {
          "type": "string"
        },
        "community_address": {
          "type": "string"
        },
        "agent_designation": {
          "type": "string"
        }
      }
    },
...
  }
}